트레이너 견적서 CRUD

+++ src/main/java/com/project/durumoongsil/teutoo/estimate/domain/Estimate.java
@@ -20,7 +20,7 @@ public class Estimate extends BaseTimeEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private Long price;
+    private Integer price;
     private Integer ptCount;
     private String ptAddress;
 
@@ -29,7 +29,7 @@ public class Estimate extends BaseTimeEntity {
     private Member member;
 
     @Builder
-    public Estimate(Long id, Long price, Integer ptCount, String ptAddress, Member member) {
+    public Estimate(Long id, int price, Integer ptCount, String ptAddress, Member member) {
         this.id = id;
         this.price = price;
         this.ptCount = ptCount;
